+++ src/game_logic/damage_infliction_system.cpp
@@ -16,7 +16,7 @@
 
 #include "damage_infliction_system.hpp"
 
-#include "data/player_data.hpp"
+#include "data/player_model.hpp"
 #include "engine/base_components.hpp"
 #include "engine/physical_components.hpp"
 #include "engine/visual_components.hpp"
