+++ src/game_logic/ingame_systems.cpp
@@ -16,7 +16,7 @@
 
 #include "ingame_systems.hpp"
 
-#include "data/player_data.hpp"
+#include "data/player_model.hpp"
 #include "engine/random_number_generator.hpp"
 #include "game_logic/entity_factory.hpp"
 #include "game_logic/interaction/force_field.hpp"
