+++ src/game_logic/player_interaction_system.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include "base/warnings.hpp"
-#include "data/player_data.hpp"
+#include "data/player_model.hpp"
 #include "engine/base_components.hpp"
 #include "game_logic/player/components.hpp"
